pds-wadmin-modules/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wadmin"
version = "2.0.0"
description = "W Social PDS Admin CLI Tool"
requires-python = ">=3.8"
dependencies = [
    "click>=8.1.0",
    "requests>=2.31.0",
    "hvac>=2.0.0",
    "rich>=13.0.0",
    "tabulate>=0.9.0",
    "python-dateutil>=2.8.0",
    "sib-api-v3-sdk>=7.6.0",
]

[project.scripts]
wadmin = "wadmin.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["wadmin*"]
